feat(courses): update course management features
* Refactor course list component to use button links for actions.
* Enhance course form validation with dynamic name length.
* Improve error handling in course resolver and service.
* Update lesson model to use Long for IDs.
* Modify course repository and service to limit search results.
* Adjust tests to reflect changes in data types and validation.
* Remove unused validation configuration class.
* Update application properties for error handling.
* Refactor YouTube URL validation to only accept video IDs.

+++ crud-spring/src/main/java/com/loiane/course/dto/LessonDTO.java
@@ -9,9 +9,10 @@ import jakarta.validation.constraints.NotNull;
 
 /**
  * Used as response and request object that represents a Lesson.
+ * On requests, a null or zero {@code _id} means a new lesson.
  */
 public record LessonDTO(
-        int _id,
+        Long _id,
         @NotBlank @NotNull @Length(min = 5, max = 100) String name,
         @NotBlank @NotNull @Length(min = 10, max = 11) @ValidYouTubeUrl String youtubeUrl) {
 }
